Version 1.8 Beta
Fixed events not being deleted even though the "Delete One Time Events" flag was set.

For afdn_countdownTimer(), $eventLimit default is now -1 instead of 0. This allows zero events to be shown (not that it makes any sense, but you can do that now).

For afdn_countdownTimer(), if $eventLimit is set and and the "Time Since" flag is not set, events would be skipped. This has been fixed.



git-svn-id: https://plugins.svn.wordpress.org/countdown-timer/trunk@12330 b8457f37-d9ea-0310-8a92-e5e31aec5664